requirements.txt
Django==5.2.1
djangorestframework==3.16.0
django-cors-headers==4.7.0
drf-spectacular==0.28.0
Pillow==10.2.0
PyYAML==6.0.2
uritemplate==4.1.1
jsonschema==4.24.0
inflection==0.5.1 